Restructured 'core' folder root files.
* Moved superdesk.docs.core module to the UI guide and out of the core
* Deleted README.md inside 'core' folder (this will now be described in ngdoc files
* Moved `superdesk.config` module to 'core/config.js' file
* Created a new 'filters' folder to contain 'core/filters.js' contents
* Moved contents of 'core/gettext.js' to 'core/lang/index.js'
* Moved code that creates `superdesk.core` module to 'core/index.js'
* Moved `superdesk.core.loading` module to its own file in 'core/loading.js'
* Moved 'core/mocks.js' to 'core/tests/mocks.js' and updated import paths
