+++ scripts/core/filters/index.js
