+++ scripts/core/tests/mocks.js
