+++ scripts/tests.js
@@ -1,6 +1,6 @@
 import 'vendor';
 import 'core';
-import 'core/mocks';
+import 'core/tests/mocks';
 import 'apps';
 
 var testsContext = require.context('.', true, /.spec.js$/);
